The large refactoring of RequestController, update of security policy.
- Split RequestController: some functionality moved to services.
- Switched to @PreAuthorize for security, as in the other project.
