+++ src/main/java/business/exceptions/ExcerptListNotFound.java
@@ -1,0 +1,11 @@
+package business.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value=HttpStatus.NOT_FOUND, reason="Excerpt list not found.")
+public class ExcerptListNotFound extends RuntimeException {
+    public ExcerptListNotFound() {
+        super("Excerpt list not found.");
+    }
+}
