+++ src/main/java/business/services/RequestService.java
@@ -1,0 +1,144 @@
+package business.services;
+
+import java.util.List;
+
+import org.activiti.engine.FormService;
+import org.activiti.engine.HistoryService;
+import org.activiti.engine.IdentityService;
+import org.activiti.engine.RuntimeService;
+import org.activiti.engine.TaskService;
+import org.activiti.engine.history.HistoricTaskInstance;
+import org.activiti.engine.runtime.ProcessInstance;
+import org.activiti.engine.task.Task;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import business.exceptions.RequestNotFound;
+import business.exceptions.TaskNotFound;
+
+@Service
+public class RequestService {
+
+    Log log = LogFactory.getLog(getClass());
+
+    @Autowired
+    private RuntimeService runtimeService;
+
+    @Autowired
+    private TaskService taskService;
+
+    @Autowired
+    private IdentityService identityService;
+
+    @Autowired
+    private FormService formService;
+
+    @Autowired
+    private HistoryService historyService;
+    
+    /**
+     * Finds current task. Assumes that exactly one task is currently active.
+     * @param requestId
+     * @return the current task if it exists.
+     */
+    public List<HistoricTaskInstance> getHistoricTasksByRequestId(String requestId) {
+        return historyService.createHistoricTaskInstanceQuery()
+                .processInstanceId(requestId)
+                .taskDefinitionKey("dna_isolation")
+                .orderByHistoricTaskInstanceEndTime()
+                .desc()
+                .list();
+    }
+
+    /**
+     * Finds current task. Assumes that exactly one task is currently active.
+     * @param requestId
+     * @return the current task if it exists.
+     * @throws business.exceptions.TaskNotFound
+     */
+    public Task getTaskByRequestId(String requestId, String taskDefinition) {
+        Task task = taskService.createTaskQuery().processInstanceId(requestId)
+                .active()
+                .taskDefinitionKey(taskDefinition)
+                .singleResult();
+        if (task == null) {
+            throw new TaskNotFound();
+        }
+        return task;
+    }
+
+    /**
+     * Finds current task. Assumes that at most one task is currently active.
+     * @param requestId
+     * @return the current task if it exists, null otherwise.
+     */
+    public Task findTaskByRequestId(String requestId, String taskDefinition) {
+        Task task = taskService.createTaskQuery().processInstanceId(requestId)
+                .active()
+                .taskDefinitionKey(taskDefinition)
+                .singleResult();
+        return task;
+    }
+
+    /**
+     * Finds current task. Assumes that exactly one task is currently active.
+     * @param requestId
+     * @return the current task if it exists.
+     * @throws business.exceptions.TaskNotFound
+     */
+    public Task getCurrentPalgaTaskByRequestId(String requestId) {
+        Task task = findCurrentPalgaTaskByRequestId(requestId);
+        if (task == null) {
+            throw new TaskNotFound();
+        }
+        return task;
+    }
+    
+    /**
+     * Finds current task. Assumes that at most one task is currently active.
+     * @param requestId
+     * @return the current task if it exists, null otherwise.
+     */
+    public Task findCurrentPalgaTaskByRequestId(String requestId) {
+        Task task = findTaskByRequestId(requestId, "palga_request_review");
+        if (task == null) {
+            task = findTaskByRequestId(requestId, "request_approval");
+        }
+        if (task == null) {
+            task = findTaskByRequestId(requestId, "data_delivery");
+        }
+        return task;
+    }
+    
+    /**
+     * Finds request.
+     * @param processInstanceId
+     * @return the current request if it exists; null otherwise.
+     */
+    public ProcessInstance findProcessInstance(String processInstanceId) {
+        ProcessInstance instance = runtimeService.createProcessInstanceQuery()
+                .includeProcessVariables()
+                .processInstanceId(processInstanceId).singleResult();
+        return instance;
+    }
+
+    /**
+     * Finds request.
+     * @param processInstanceId
+     * @return the current request if it exists.
+     * @throws
+     */
+    public ProcessInstance getProcessInstance(String processInstanceId) {
+        ProcessInstance instance = runtimeService.createProcessInstanceQuery()
+                .includeProcessVariables()
+                .processInstanceId(processInstanceId).singleResult();
+        if (instance == null) {
+            throw new RequestNotFound();
+        }
+        return instance;
+    }
+
+    
+}
