+++ src/main/java/business/services/UserService.java
@@ -1,4 +1,4 @@
-package business.models;
+package business.services;
 
 import javax.transaction.Transactional;
 
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import business.controllers.UserController.EmailAddressNotAvailableException;
+import business.models.User;
+import business.models.UserRepository;
 
 @Service
 @Transactional
