+++ src/main/resources/application.properties
@@ -7,7 +7,7 @@ dntp.server-port = 8092
 #spring.profiles.default=dev
 
 # logging level
-logging.level. = DEBUG 
+logging.config = logback.groovy
 
 # Hibernate settings
 spring.jpa.generate-ddl = true
