+++ src/main/resources/static/app/request/editrequest.html
@@ -138,7 +138,7 @@
                     </div>
 
                     <!-- Informed Consent  -->
-                    <div class="form-group" ng-class="" ng-show="request.linkageWithPersonalData == 'true'">
+                    <div class="form-group" ng-class="" ng-show="request.linkageWithPersonalData == true">
                         <label class="control-label col-sm-3" for="informedConsentYes">
                             {{'Informed Consent' | translate}}
                         </label>
